==> Hulu_and_Disney/01_get_tv_shows_hulu_disney.ipynb <==
import requests
import bs4
import pandas as pd
import numpy as np

## Get TV shows from Hulu and Disney+

#### Goal

- Use webscraping to get the tv shows
- Export each data frame as csv

#### General information:

- Hulu has 1,744 tv shows, 50 per page, 35 pages in total.

- Disney+ has 179 tv shows, 50 per page, 4 in total.


### 1. Define function to get tv shows and additional information

We will use webscarping to get the tv shows from the website reelgood.com.

Each page has up to 50 tv shows.

**NOTE: We are using the same function as in 01_get_tv_shows_df notebook.**

# Define function that takes the url and number of tv shows you want to get from it, default is 50 which is the maximum.

def get_tv_shows(url, no=50):
    
    # define empty variables
    shows = {}
    year = []
    rating = []
    imdb = []
    rotten_tom = []
    
    # get response and contents using beautiful soup
    resp = requests.get(url)
    soup = bs4.BeautifulSoup(resp.content, "html.parser")
    tv_titles = soup.find("table", class_="css-1179hly").find_all("td", class_="css-1u7zfla")
    extras = soup.find("table", class_="css-1179hly").find_all("td", class_="css-1u7zfla")
    
    # get list of tv shows
    shows["show"] = [tv_titles[i].find("a").contents[0] for i in range(no)]
    
    # get lists for years, rating, imdb and rotten tomatoes reviews
    for i in range(no):
        year.append(soup.find("table", class_="css-1179hly").find_all("td", class_="css-1u11l3y")[i*4+0].contents[0])
        rating.append(soup.find("table", class_="css-1179hly").find_all("td", class_="css-1u11l3y")[i*4+1].contents[0])
        imdb.append(soup.find("table", class_="css-1179hly").find_all("td", class_="css-1u11l3y")[i*4+2].contents[0])
        rotten_tom.append(soup.find("table", class_="css-1179hly").find_all("td", class_="css-1u11l3y")[i*4+3].contents[0])

    shows["year"] = year
    shows["rating"] = rating
    shows["imdb"] = imdb
    shows["rotten_tomatoes"] = rotten_tom
    
    return shows

## 2. Get tv shows for each streaming service using get_tv_shows function


## Hulu

url = "https://reelgood.com/tv/source/hulu"
pages = 0
hulu_shows = pd.DataFrame(columns=["show", "year", "rating", "imdb", "rotten_tomatoes"])

while pages < 1700:
    new_pd = pd.DataFrame(get_tv_shows(url))
    hulu_shows = hulu_shows.append(new_pd)
    pages +=50
    url = "https://reelgood.com/tv/source/hulu?offset=" + str(pages)

hulu_shows = hulu_shows.append(pd.DataFrame(get_tv_shows("https://reelgood.com/tv/source/hulu?offset=1700", 44)))


hulu_shows.head()

## Disney+

url = "https://reelgood.com/tv/source/disney_plus"
pages = 0
disney_shows = pd.DataFrame(columns=["show", "year", "rating", "imdb", "rotten_tomatoes"])

while pages < 150:
    new_pd = pd.DataFrame(get_tv_shows(url))
    disney_shows = disney_shows.append(new_pd)
    pages +=50
    url = "https://reelgood.com/tv/source/disney_plus?offset=" + str(pages)

disney_shows = disney_shows.append(pd.DataFrame(get_tv_shows("https://reelgood.com/tv/source/disney_plus?offset=150", 29)))


disney_shows.head()

## 3. Export data

# hulu_shows.to_csv("Data_Hulu_Disney/hulu_shows.csv", index_label=False)
# disney_shows.to_csv("Data_Hulu_Disney/disney_shows.csv", index_label=False)
