==> Hulu_and_Disney/02.1_b_disney_imdb_merge_and_clean.ipynb <==
import pandas as pd
import numpy as np
from fuzzywuzzy import fuzz

## Hulu Merge

#### Goals:

- Merge Hulu tv shows with IMDB ID


## 1. Import imdb and hulu data

imdb = pd.read_pickle("../Data/imdb_tv_all.pkl")

disney = pd.read_csv("Data_Hulu_Disney/disney_shows.csv")

We will look at how many titles we have using the shape function, which will return a tuple indicating number of rows and number of columns.

disney.shape

## 2. Merge dataframes

We will start by merging on title and year in order to get the right IMDB ID (tconst) because there might be some remakes.
We will do this by using a left join in order to keep all the original hulu titles.

disney_genres = disney.merge(imdb, left_on=["show", "year"], right_on=["originalTitle", "startYear"], how="left")

From the new merged data frame, we will get a new data frame called hulu_missing1 to get the missing titles that didn't find a match.

disney_missing1 = disney_genres[disney_genres["tconst"].isna()].reset_index(drop=True)[["show", "year", "rating"
                                                                                           , "imdb", "rotten_tomatoes"]]

disney_missing1.shape

We are missing 41 titles from 179 in total.

We will now merge but now using primaryTitle instead of originalTitle, since some of the titles might be in their original language.

disney_missing1 = disney_missing1.merge(imdb, left_on=["show", "year"], right_on=["primaryTitle", "startYear"], how="left")

We will create another data frame for those titles who weren't matched.

disney_missing2 = disney_missing1[disney_missing1["tconst"].isna()].reset_index(drop=True)[["show", "year", "rating"
                                                                                               , "imdb", "rotten_tomatoes"]]

disney_missing2.shape

We are now missing 39 titles. Since we don't have something else to relate to in this dataframe, we will now import the other data frame from the IMDB data base which includes all different titles created for a single one and the IMDB ID (titleID).

title_regions = pd.read_csv("C:/Users/Admin/Documents/ironhack/title.akas.tsv.gz", sep="\t", low_memory=False)

title_regions.head()

We will do the merge of the missing titles with the new title_regions data frame by title name.

disney_missing2 = disney_missing2.merge(title_regions, left_on=["show"], right_on=["title"], how="left")

disney_missing3 = disney_missing2[disney_missing2["titleId"].isna()].reset_index(drop=True)[["show", "year", "rating"
                                                                                                , "imdb", "rotten_tomatoes"]]

disney_missing3.shape

Now we are just missing 21 titles. We will now try to find the missing title IDs with fuzzy wuzzy.

### 3. Find missing titles using fuzzywuzzy

We will first create a list for missing shows.

disney_shows_missing = [show for show in disney_missing3["show"]]

imdb_titles = [title for title in imdb["primaryTitle"]]

We will use the same function we created on 03.2_a_netflix_missing_title_ids to find the most similar title on the imdb title basics file.

def find_shows(show):
    matches = []

    for title in imdb_titles:
        # compute ratio and remove case-sensitivity
        ratio = fuzz.ratio(title.lower(), show.lower())

        # add all matches to list with ratio > 60
        if ratio >= 60:
            matches.append((title, show, ratio))
    
    # return none if there was no match found
    if len(matches) == 0:
        return None
    return sorted(matches, key=lambda x: x[2], reverse=True)[0][0]

disney_missing3["imdb_titles"] = disney_missing3["show"].apply(lambda x: find_shows(x))

disney_missing3.head(10)

From the first 10 missing titles we can see that our function did a good job, we will now take a look at the titles that weren't found a match.

disney_missing3[disney_missing3["imdb_titles"].isna()]

We found a match for all the missing titles, we will now merge them with their IMDB ID and the year.

## 4. Merge IMDB IDs with missing titles

We will still merge on the year, in order to be sure that we are not merging a remake or a similar title.

disney_missing3 = disney_missing3.merge(imdb, how="left", left_on=["imdb_titles", "year"], right_on=["primaryTitle", "startYear"])

disney_missing3.head()

We will now take a look at the titles who didn't find a match.

disney_missing3[disney_missing3["tconst"].isna()]

We can see that most of this titles don't have a similar imdb titles, due to this, we will drop this rows.

disney_missing_ids = disney_missing3[~disney_missing3["tconst"].isna()].reset_index(drop=True).drop_duplicates("tconst")

Now we are just missing 10 titles. We will remove this titles for our project.

## 5. Join data frames with IMDB IDs

We will join all data frames without null values to get a final hulu data frame containing the IMDB ID.
We will drop duplicates because the title regions may have joined multiple rows since some titles have the same name in different regions.

disney_genres = disney_genres[~disney_genres["tconst"].isna()]

disney_genres = disney_genres.append(disney_missing1[~disney_missing1["tconst"].isna()])

disney_genres = disney_genres.append(disney_missing2[~disney_missing2["titleId"].isna()])

disney_genres = disney_genres.append(disney_missing_ids)

disney_genres = disney_genres.drop_duplicates("tconst").reset_index(drop=True)

disney_genres.head()

We will now calculate the ratio of titles with an imdb id match and the original titles.

len(disney_genres) / len(disney)

This means that we will be using 85% of the original data.

## 6. Cleaning final data

We will create a final data frame including:
- show
- year
- rating
- imdb
- rotten_tomatoes
- imdb_id
- all data from title basics

First, we will need to create the imdb_id column, this will have tconst and titleId values.

disney_genres["imdb_id"] = np.where(disney_genres["tconst"].isna(), disney_genres["titleId"], disney_genres["tconst"])

disney_genres["imdb_id"].isna().value_counts()

This means we have all the IMDB IDs for all the titles in this final data frame. We will now remove columns that are not show, year, rating, imdb, rotten_tomatoes and imdb_id and merge again.

to_drop = [col for col in disney_genres.columns if col not in ["show", "year", "rating", "imdb", "rotten_tomatoes", "imdb_id"]]

disney_genres = disney_genres.drop(columns=to_drop)

disney_genres.head()

Now we will merge all data from title basics

disney_genres = disney_genres.merge(imdb, how="left", left_on="imdb_id", right_on="tconst")

disney_genres.head()

disney_genres = disney_genres.drop(columns="tconst")

## 7. Check null and unique values

disney_genres.isna().sum()

disney_genres.nunique(axis=0)

From the previous cells we can see that:
- Most values are missing for rotten_tomatoes.
- isAdult has just one value.

We will drop this two columns, since we cannot get much information from it.

disney_genres = disney_genres.drop(columns=["rotten_tomatoes", "isAdult"])

## 8. Change data types

- rating: We will remove the '+' sign and turn it into an integer
- runtimeMinutes: we will change the type to integer

#### i. rating

disney_genres["rating"] = [str(i).replace("+", "") for i in disney_genres["rating"]]

disney_genres["rating"].value_counts()

Since rating is a string type, we will convert the 'nan' values to null and 'all' to 0, meaning that the series can be watched by all ages.

disney_genres["rating"] = np.where(disney_genres["rating"] == "nan", None, disney_genres["rating"])
disney_genres["rating"] = np.where(disney_genres["rating"] == "all", 0, disney_genres["rating"])

disney_genres["rating"].value_counts()

disney_genres["rating"] = pd.to_numeric(disney_genres["rating"], errors="coerce")

#### ii. runtimeMinutes

disney_genres["runtimeMinutes"] = pd.to_numeric(disney_genres["runtimeMinutes"], errors="coerce")

#### ii. Check final data types

disney_genres.dtypes

## 9. Rename rating and imdb columns

disney_genres = disney_genres.rename(columns={"rating":"age", "imdb":"imdb_rating"})

## 10. Export data

# disney_genres.to_pickle("Data_Hulu_Disney/disney_final_clean.pkl")
