==> Hulu_and_Disney/05.1_streaming_service_recommender_df_(with_Hulu_and_Disney).ipynb <==
import pandas as pd
import numpy as np

pd.set_option("display.max_columns", None)

## Streaming Service Recommender
#### Goals

- Build data frame to be used for recommender

We want our final data frame to llok like the following:

| streaming_service | Drama | Comedy | Animation | ... |
|-------------------|-------|--------|-----------|-----|
| **Netflix**       | 0.3765 | 0.2735 | 0.1680   | ....|
| **Amazon**        | 0.3349 | 0.2432 | 0.1619   | ... |
| ... | ... | ... |... | ...|

Our genres should be our columns and the streaming services should be the index.

### 1. Import data

genres_recommender = pd.read_pickle("../Data/genres_recommender.pkl")

hulu = pd.read_pickle("Data_Hulu_Disney/hulu_genres_ratio.pkl")

disney = pd.read_pickle("Data_Hulu_Disney/disney_genres_ratio.pkl")

### 2. Get list of total genres

We will first get the lists of genres for each streaming service, we will them add them together and afterwards convert it to a set to remove duplicates.

genres = genres_recommender.columns.to_list()

hulu_genres = hulu["genre"].to_list()

disney_genres = disney["genre"].to_list()

total_genres = list(set(genres + hulu_genres + disney_genres))


total_genres.sort()

len(total_genres)

total_genres

### 3. Create new data frame

#### i. Prepare each data frame

hulu["streaming_service"] = "Hulu"

hulu_genres = hulu[["genre", "ratio", "streaming_service"]]

hulu_genres.head()

disney["streaming_service"] = "Disney"

disney_genres = disney[["genre", "ratio", "streaming_service"]]

disney_genres.head()

#### ii. Append data frames

genres_features = hulu_genres.append(disney_genres).reset_index()

#### iii. Create pivot table

genres_recommender_2 = genres_features.pivot_table(index="streaming_service",
                                      columns="genre",
                                      values="ratio").fillna(0)

genres_recommender_2

We will now append genres_recommender_2 to th genres_recommender data frame, which has Netflix, Amazon and HBO's genre information.

genres_recommender = genres_recommender.append(genres_recommender_2)

genres_recommender

### 4. Export final data frame

# genres_recommender.to_pickle("Data_Hulu_Disney/genres_recommender_v2.pkl")

import pandas as pd
df = pd.read_pickle("../Data/netflix_genres.pkl")

df[df["Horror"] == 1]
