==> Hulu_and_Disney/05.3_straming_service_recommender_model.ipynb <==
import pandas as pd
import numpy as np
import re 
from fuzzywuzzy import fuzz
from scipy.spatial.distance import pdist, squareform

pd.set_option("display.max_columns", None)

## Streaming Service Recommender Model

#### Goals

- Test the streaming service recommender model created previously on 06.3_streaming_service_recommender_model notebook

### 1. Import data


genres_recommender = pd.read_pickle("Data_Hulu_Disney/genres_recommender_v2.pkl")

genres_recommender

### 2. Define function get_streaming_recommendation

def get_streaming_recommendation(genres_recommender, user_survey):
    
    recommendations = []
    
    user_recommender = genres_recommender.append(user_survey)
        
    distances = squareform(pdist(user_recommender, 'euclidean'))
    
    distances_df = pd.DataFrame((squareform(pdist(user_recommender, 'euclidean'))), 
                         index=user_recommender.index, columns=user_recommender.index)
    
    similar_streamings = list(distances_df['User'].sort_values()[1:].index)
    distances = list(distances_df['User'].sort_values()[1:].values)
    
    # we will increase the range in order to get the similarities for all 5 streaming services
    for i in range(5):
        recommendations.append(similar_streamings[i])
       
    # we will show the percentage of similarity, whcih is 1 minus the distance multiplied by a 100
    print("Your closest match is " + recommendations[0] + f" with {int(100-distances[0]*100)}% genre similarity.")
    print("--------------------------")
    print("Your next matches are:")
    print("2. " + recommendations[1] + f" with {int(100-distances[1]*100)}% genre similarity.")
    print("3. " + recommendations[2] + f" with {int(100-distances[2]*100)}% genre similarity.")
    print("4. " + recommendations[3] + f" with {int(100-distances[3]*100)}% genre similarity.")
    print("5. " + recommendations[4] + f" with {int(100-distances[4]*100)}% genre similarity.")
    
    return ("")


### 3. Test

user_netflix = pd.read_pickle("../Data/netflix_test.pkl")
user_amazon = pd.read_pickle("../Data/amazon_test.pkl")
user_hbo = pd.read_pickle("../Data/hbo_test.pkl")
user_hulu = pd.read_pickle("Data_Hulu_Disney/hulu_test.pkl")
user_disney = pd.read_pickle("Data_Hulu_Disney/disney_test.pkl")

get_streaming_recommendation(genres_recommender, user_netflix)

get_streaming_recommendation(genres_recommender, user_amazon)

get_streaming_recommendation(genres_recommender, user_hbo)

get_streaming_recommendation(genres_recommender, user_hulu)

get_streaming_recommendation(genres_recommender, user_disney)

The survey worked, we can now create a module and notebook for the new data frames including Hulu and Disney+.
