==> Hulu_and_Disney/05.4_survey_shows.ipynb <==
import pandas as pd

pd.set_option("display.max_columns", None)

# Survey Shows Data Frame

#### Goals

- We will create a data frame of genres for the shows that are in the survey in order to use this data frame instead of the data frame of all shows.

### 1. Import data

all_shows = pd.read_pickle("Data_Hulu_Disney/all_shows_genres_v2.pkl")

### 2. Create list of shows in survey

survey_shows_lst = ["revisions", "Survivor", "John Adams", "Rick and Morty", "Sonny with a Chance", "Big Mouth", "Sharp Objects", "Violetta",
                    "The Amazing Race", "The Son", "Klondike", "Fill in the Prank", "The Sopranos", "American Horror Story",
                    "One Strange Rock", "The Mandalorian", "The Planets", "Chilling Adventures of Sabrina", "Vinyl", "Zoey's Extraordinary Playlist",
                     "Ballers", "The Joy of Painting", "Good Morning America", "Boy Meets World", "Hart of Dixie", "Black Mirror",
                    "The World According to Jeff Goldblum", "Eat. Race. Win", "What We Do in the Shadows", "Any Given Wednesday with Bill Simmons",
                     "Stranger Things", "Big Hero 6 The Series", "Silicon Valley", "Generation Kill", "Star Trek", "The Night Of", "Bonanza",
                     "X-Men: Evolution", "Crazy Ex-Girlfriend", "Atlanta", "How The Universe Works", "Last Week Tonight with John Oliver",
                     "Supernatural", "How I Met Your Mother", "Brain Games", "Strictly Ballet", "VICE News Tonight", "Buffy the Vampire Slayer",
                     "Star Wars: The Clone Wars", "Fauda", "Baseball", "The Witcher", "Pixar in Real Life", "Westworld", "Bob's Burgers",
                     "Cheer", "DuckTales", "The Leftovers", "The Boys", "MythBusters", "Ozark", "Teen Wolf", "Killing Eve", "So Weird",
                     "Flight of the Conchords", "Barry", "M*A*S*H", "Tiger King: Murder, Mayhem and Madness", "Chuck", "Doc McStuffins", 
                     "Yeh Meri Family", "Watchmen", "Psycho-Pass", "RuPaul's Drag Race", "Life Below Zero", "The Durrells", "Recess",
                     "His Dark Materials", "Five Came Back", "Brooklyn Nine-Nine", "The Circle", "Lab Rats", "Veep", "Fortitude", "Twin Peaks",
                     "Girls", "High School Musical: The Musical: The Series", "Mozart in the Jungle", "Rythm + Flow", "The Tonight's Show Starring Jimmy Fallon",
                     "Curb Your Enthusiasm", "The Young Turks", "Shark Tank", "The 100", "Hannah Montana", "The Man in the High Castle",
                     "The Motorbike Show", "Run", "Archer", "Marvel's Agent Carter", "Don't Watch This", "Modern Love", "Friday Night Lights",
                     "Lizzie McGuire", "Angels in America", "Hard Knocks", "Blue Mountain State", "The Rifleman", "Castle Rock", "America's Funniest Home Videos",
                     "The Curious CReations of Christine McConnell", "Folklore", "Just Add Magic", "Fargo", "Phineas and Ferb",
                     "The Hitchhiker's Guide to the Galaxy", "JONAS", "The Vietnam War", "Animals.", "Happy Endings", "Hell's Kitchen",
                     "Malcolm in the Middle", "The Defiant Ones", "The Imagineering Story", "The Ranch", "The End of the F***ing World",
                     "The Jinx: The Life and Deaths of Robert Durst", "Horrible Histories", "The Handmaid's Tale", "K.C. Undercover", 
                     "Togetherness", "Poldark", "Once Upon a Time", "Genius", "Encore!", "Destino Rusia", "The Flash", "That's So Raven",
                     "Good Omens", "Justified", "Undone", "Deadwood", "BoJack Horseman", "Parks and Recreation", "The Simpsons", 
                     "Love, Death & Robots", "Big Little Lies", "The Snowy Day", "Top Chef", "Running Wild with Bear Grylls"]

### 3. Create new data frame and export data

survey_shows = all_shows[all_shows["show"].isin(survey_shows_lst)].reset_index(drop=True)

survey_shows.head()

# survey_shows.to_pickle("Data_Hulu_Disney/survey_shows.pkl")
